requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
python-dotenv==0.20.0
openai==0.19.0
aiofiles==0.8.0